requirements.txt
et-xmlfile==1.1.0
future==0.18.2
greenlet==1.0.0
iso8601==0.1.14
numpy==1.20.3
pandas==1.2.4
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.13
xlrd==2.0.1
XlsxWriter==1.4.3
